requirements.txt
# Core Libraries
python-dotenv
streamlit>=1.30.0
pandas>=2.0.0
requests
streamlit-extras>=0.2.0
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.2
plotly>=5.17.0
numpy>=1.24.0
scikit-learn>=1.2.2
plotly>=5.17.0
